lab-security/requirements.txt
dotenv
azure-ai-inference
opentelemetry-sdk